+++ src/handlers/ai/a2a/cancelTask.ts
@@ -1,6 +1,7 @@
 import log from 'electron-log/main'
 
 import { LOG_PREFIX } from './constants'
+import { buildA2AHeaders } from './helpers'
 import type { A2ASessionConfig } from './types'
 
 export async function sendTaskCancel(
@@ -16,11 +17,7 @@ export async function sendTaskCancel(
 
   const response = await fetch(`${config.baseUrl}/agents/${config.agentId}`, {
     method: 'POST',
-    headers: {
-      'Content-Type': 'application/json',
-      Authorization: `Bearer ${config.bearerToken}`,
-      'X-App-Source': 'k6-studio',
-    },
+    headers: buildA2AHeaders(config),
     body: JSON.stringify(body),
   })
