+++ src/handlers/ai/a2a/constants.ts
@@ -1,1 +1,7 @@
 export const LOG_PREFIX = '[GrafanaAssistant]'
+
+export const EMPTY_USAGE = {
+  inputTokens: undefined,
+  outputTokens: undefined,
+  totalTokens: undefined,
+} as const
