+++ src/views/Generator/AutoCorrelation/AutoCorrelation.tsx
@@ -124,19 +124,15 @@ export function AutoCorrelation({
 
       <Flex
         gap="3"
-        justify="between"
+        justify={provider === 'openai' ? 'between' : 'end'}
         align="center"
         p="3"
         css={{
           borderTop: '1px solid var(--gray-5)',
         }}
       >
-        {provider === 'openai' ? (
-          <div>
-            <TokenUsageIndicator tokenUsage={tokenUsage} />
-          </div>
-        ) : (
-          <div />
+        {provider === 'openai' && (
+          <TokenUsageIndicator tokenUsage={tokenUsage} />
         )}
         <Flex gap="3">
           <Button
